feat: add proxy module with createProxyApp function and update exports
