+++ packages/cli/package.json
@@ -37,7 +37,20 @@
         "pack": "pnpm pack"
     },
     "exports": {
-        "./package.json": "./package.json"
+        "./package.json": {
+            "import": "./package.json",
+            "require": "./package.json"
+        },
+        "./proxy": {
+            "import": {
+                "types": "./dist/proxy.d.mts",
+                "default": "./dist/proxy.mjs"
+            },
+            "require": {
+                "types": "./dist/proxy.d.cts",
+                "default": "./dist/proxy.cjs"
+            }
+        }
     },
     "dependencies": {
         "@zenstackhq/common-helpers": "workspace:*",
